// src/student/dto/students.dto.ts
class StudentData{
    sname:String;
    teacher:String;
}
export class FindStudentsResponseDto {
    id: string;
    name: string;
    teacher: string
}

export class CreateStudentDto {
    name: string;
    teacher: string
}

export class StudentResponseDto {
    id: string;
    name: string;
    teacher: string
}

export class UpdateStudentDto {
    name: string;
    teacher: string
}